TarsOnlineTools/requirements.txt
scapy==2.4.5
speedtest-cli==2.1.3
colorama==0.4.4
rich==10.12.0
requests==2.26.0
netifaces==0.11.0
ipaddress==1.0.23